refactor(scrollview): split into subdirectories
